move towards allowing multiple torrent listeners

+++ shallows/cmd/retrovibed/daemons/cmd.daemon.go
@@ -61,7 +61,7 @@ type Command struct {
 	AutoArchive          bool             `flag:"" name:"auto-archive" help:"EXPERIMENTAL: enable automatic archiving of eligible media" default:"false" env:"${env_auto_archive}"`
 	HTTP                 cmdopts.Listener `flag:"" name:"http-address" help:"address to serve daemon api from" default:"tcp://:9998" env:"${env_daemon_socket}"`
 	SelfSignedHosts      []string         `flag:"" name:"self-signed-hosts" help:"comma seperated list of hosts to add to the sign signed certificate" env:"${env_self_signed_hosts}"`
-	TorrentPort          int              `flag:"" name:"torrent-port" help:"port to use for torrenting" env:"${env_torrent_port}" default:"10000"`
+	TorrentPort          uint16           `flag:"" name:"torrent-port" help:"port to use for torrenting" env:"${env_torrent_port}" default:"10000"`
 	TorrentPrivate       bool             `flag:"" name:"torrent-private" help:"restrict torrent connections to private networks" env:"${env_torrent_private}" default:"false"`
 	TorrentPublicIP4     string           `flag:"" name:"torrent-ipv4" help:"public ipv4 address of the torrent" env:"${env_torrent_ipv4}"`
 	TorrentPublicIP6     string           `flag:"" name:"torrent-ipv6" help:"public ipv6 address of the torrent" env:"${env_torrent_ipv6}"`
